% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achd_data.R
\docType{data}
\name{ACHD}
\alias{ACHD}
\title{Australian Coronary Heart Disease Study (1991)}
\format{
An object of class \code{"data.table"}
}
\usage{
data(ACHD)
}
\description{
Heart disease study. Age group index for 5-year categories
(i = 1 represents 30-34, i = 2 represents 35-39, ..., i = 8 represents 65-69).
The data represents the number of males in age group i who died from coronoary
heart disease in the Hunter region of New South Wales (1991) along with the
population of that age group.
}
\examples{
data(ACHD)


}
\references{
BIOS:7410 class notes
}
\keyword{datasets}
