% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact.R
\name{compact}
\alias{compact}
\title{Compact Function}
\usage{
compact(x)
}
\arguments{
\item{x}{A list}
}
\description{
Removes NULL entrires from list (stolen from plyr)
}
\examples{
l <- vector("list", length = 5)
l[[1]]] <- "A"
print(l)
l <- compact(l)
print(l)
}
