% Generated by roxygen2 (4.0.1): do not edit by hand
\name{content}
\alias{content}
\title{Summarize All Variables in a Data Frame}
\usage{
content(data, sep = ",", class = TRUE, print.width = getOption("width"),
  print = TRUE)
}
\arguments{
\item{data}{a data frame to reorder}

\item{sep}{a character string to separate printed values for a variable}

\item{class}{logical: whether to display the class of each variable}

\item{print.width}{an integer indicating the max number of characters to
print on a line}

\item{print}{logical: whether to print anything}
}
\value{
Invisibly returns a data frame with the summary.  This has special
attributes "nrow", "ncol", "nrow.miss", and "ncol.miss", containing the
number of rows, columns, rows with missing values, and columns with
missing values in \code{data}, respectively.
}
\description{
This function prints an easy-to-read summary of a data frame.  Each
variable takes up only one line of output, allowing much information
to be quickly gleaned.  This is similar to \code{\link[utils]{str}}.
}
\details{
The output starts by listing the number of variables and observations. It then
displays a table with one row per variable. The 1st column is the variable name,
optionally followed by the class, and then an abbreviated list of values.
}
\examples{
content(ChickWeight)
}
\author{
Daniel Muenz
}
\seealso{
\code{\link[utils]{str}}, \code{\link{summary}}
}
\keyword{documentation}
\keyword{print}
\keyword{utilities}

