% Generated by roxygen2 (4.0.1): do not edit by hand
\name{freq}
\alias{as.data.frame.freq}
\alias{dim.freq}
\alias{freq}
\alias{print.freq}
\title{Frequency and Percent Tabulations for Data Frames and Vectors}
\usage{
freq(..., cross = FALSE, useNA = FALSE, order = c("value", "freq"))

\method{as.data.frame}{freq}(x, ...)

\method{dim}{freq}(x)

\method{print}{freq}(x, decimals = 2, big.mark = ",", ...)
}
\arguments{
\item{\dots}{variables to tabulate; can contain data.frames and vectors.
If \code{cross==TRUE}, then variables must all have the same length.
(For S3 methods, \dots is unused.)}

\item{cross}{logical: whether to cross-tabulate variables. Default is FALSE.}

\item{useNA}{TRUE/FALSE/NULL: how to include NA's in table. FALSE (default) says
display NA's but ignore them in calculating percents; TRUE says use
them in calculating percents; and NULL says ignore them entirely}

\item{order}{"value"/"freq": controls order of freq tables. "value" (default)
sorts by variable values; "freq" sorts by descending frequency.
You can specify just the initial letter.}

\item{x}{an object of class "freq", as returned by \code{freq}.}

\item{decimals}{integer indicating the number of decimal places to be used when
printing percentages.}

\item{big.mark}{character: if not empty, used as mark between every 3 digits
  before the decimal point, when printing frequencies.}
}
\value{
\code{freq} returns an object of class "freq". If \code{cross=TRUE}, the
object is essentially a data frame. It will contain a column for each variable
being cross-tabulated and four additional columns for the tabulations.  If
\code{cross=FALSE}, the returned object will contain a list of data frames,
one per variable.

\code{as.data.frame} returns a single data frame. If \code{cross=FALSE} was
used, so that the "freq" object contains a list of data frames, then they are
vertically concatenated.
}
\description{
\code{freq} performs (cross-)tabulations for all the data frames
and vectors passed to \dots.  By default, each variable is tabulated
separately; use \code{cross=TRUE} for cross-tabulations.  Unlike with
\code{table}, \code{freq} returns frequencies, percents, and cumulative
frequencies and percents (similar to PROC FREQ in SAS).
}
\details{
\code{freq} is basically a wrapper around \code{table}. It was inspired by PROC
FREQ in SAS, and similarly displays frequences, percents, and their cumulative
cousins.  It's useful for quickly assessing the content of a data frame or a set
of variables.
}
\examples{
# sample from 1:10 and look at the distribution
freq(x=sample(1:10,200,replace=TRUE))

# make a data frame and vector for use below
n <- 150
d <- data.frame(x=sample(2,n,TRUE), y=sample(letters[1:3],n,TRUE))
is.na(d$x[sample(nrow(d),20)]) <- TRUE
is.na(d$y[sample(nrow(d),10)]) <- TRUE
z <- sample(c("i","ii","iii","iv"),n,TRUE)

freq(d) # gets freqs for all vars in d
freq(d, z) # all vars in d and z

freq(d, cross=TRUE, useNA=NULL) # coss-tab the vars in d, ignoring NAs

as.data.frame(freq(d)) # combine output from univariate freq into 1 data frame
}
\author{
Daniel Muenz
}
\seealso{
\code{\link{table}}, \code{\link{summary}}
}
\keyword{documentation}
\keyword{print}
\keyword{utilities}

