% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cell_chisq.R
\name{get_cell_chisq}
\alias{get_cell_chisq}
\title{Chi-Square of Cell Counts}
\usage{
get_cell_chisq(x, type = "Pearson")
}
\arguments{
\item{x}{A contingency table}

\item{type}{character vector indicating chi-square statistic to use. Default
is "Pearson", though "LR" and "adjusted" are also available}
}
\description{
Returns the chi-square values of cell counts in a contingency table
}
