% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HA_data.R
\docType{data}
\name{HA}
\alias{HA}
\title{Asprin and Heart Attack Study}
\format{
An object of class \code{"data.table"}
}
\usage{
data(HA)
}
\description{
Asprin and Heart Attack Study
}
\examples{
data(HA)
print(HA)
tabs <- proc_freq(Hawkeye, "Freq", "Trt_Grp*MI / RelRisk")
}
\references{
BIOS:7410 class notes
}
\keyword{datasets}
