% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkeye_data.R
\docType{data}
\name{Hawkeye}
\alias{Hawkeye}
\title{Hawkeye Football Annual Win/Loss Record}
\format{
An object of class \code{"data.table"}
}
\usage{
data(Hawkeye)
}
\description{
Hawkeye Football Annual Win/Loss Record
}
\examples{
data(Hawkeye)
print(Hawkeye)
tabs <- proc_freq(Hawkeye, "Freq", "Year*Win_Loss / Measures")
}
\references{
BIOS:7410 class notes
}
\keyword{datasets}
