% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_chisq_table.R
\name{make_chisq_table}
\alias{make_chisq_table}
\title{SAS Chi-Square Table}
\usage{
make_chisq_table(x, exact = FALSE)
}
\arguments{
\item{x}{A contingency table}

\item{exact}{Boolean argument, default FALSE, to use Fisher's e xact test
for tables larger than 2x2}
}
\description{
This function emulates the 'ChiSq' table returned by SAS
}
\examples{
A <- structure(1:4, dim = c(2L, 2L), 
               dimnames = list(c("A", "B"), c("X", "Y")))
make_chisq_table(A)
}
