% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_measures_table.R
\name{make_measures_table}
\alias{make_measures_table}
\title{This function emulates the 'Measures' table returned by SAS}
\usage{
make_measures_table(x)
}
\arguments{
\item{x}{A contingency table}
}
\description{
This function emulates the 'Measures' table returned by SAS
}
\examples{
A <- structure(1:4, dim = c(2L, 2L), 
               dimnames = list(c("A", "B"), c("X", "Y")))
make_measures_table(A)
}
