% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_relrisk_table.R
\name{make_relrisk_table}
\alias{make_relrisk_table}
\title{Create SAS Relative Risk Table}
\usage{
make_relrisk_table(x)
}
\arguments{
\item{x}{A 2x2 contingency table}
}
\description{
This function emulates the RelRisk table returned by SAS
}
\examples{
A <- structure(1:4, dim = c(2L, 2L), 
               dimnames = list(c("A", "B"), c("X", "Y")))
make_relrisk_table(A)
}
