% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sas_freq_table.R
\name{make_sas_freq_table}
\alias{make_sas_freq_table}
\title{Emulate SAS Frequency Table}
\usage{
make_sas_freq_table(x, cellchi2 = FALSE)
}
\arguments{
\item{x}{This is a casted contingency table, generated from data.table::dcast}

\item{cellchi2}{Boolean value to include cell chi-square values. Default is FALSE}
}
\value{
Currently, this just returns a matrix, but when called from 
\code{proc_freq}, it has class \code{sas_table}
}
\description{
Main function of proc_freq, used to generate the var1*var2 output
tables, with the option to include cell chi-square values. This function is primarily
intended to be called within \code{proc_freq}, so no error checking is done
}
\examples{
data(Hawkeye)
ctab <- dcast(Hawkeye, Year ~ Win_Loss, value.var = "Freq")
ctab <- as.matrix(ctab[, 2:3], rownames = ctab[[1]])
make_sas_freq_table(ctab)
}
