% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maryjane_data.R
\docType{data}
\name{maryjane}
\alias{maryjane}
\title{Alcohol, Cigarettes, and Marijuana Use}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(maryjane)
}
\description{
An Analysis of Alcohol, Cigarette, and Marijuana Use Among
High-School Seniors (1992; Dayton, Ohio)
}
\examples{
data(maryjane)


}
\references{
BIOS:7410 class notes
}
\keyword{datasets}
