% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_sas_table.R
\name{print.sas_table}
\alias{print.sas_table}
\title{Printing SAS Tables}
\usage{
\method{print}{sas_table}(x, na.print = "-", digits = 4, ...)
}
\arguments{
\item{x}{Object of class \code{sas_table}}

\item{na.print}{Basically the only thing that has changed. This default
argument replaces NA values with "-"}

\item{digits}{This is set to 4 because default print is a bit much}

\item{...}{these are your standard print options, based on the default class
of the object}
}
\description{
Method to print output from proc_freq and other sassy functions
}
