% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_freq.R
\name{proc_freq}
\alias{proc_freq}
\title{(limited) Emulation of SAS proc_freq}
\usage{
proc_freq(x, weight, tables)
}
\arguments{
\item{x}{A data.frame, data.table, or matrix containg data. Tables will
be generated with variables in the order that the data is coded. Data should
be in long format (see Hawkeye data)}

\item{weight}{The variable/column name in x the represents the frequency or
count, as a length one character vector, i.e., "Freq"}

\item{tables}{This is similar to the 'TABLES' argument in SAS. This should
also be a length one character vector. At a minimum, this should contain
two variables from x that we are interested in tabulating, joined with an '*'.
Similar to SAS code, we can request a number of tables and options, including Measures,
RelRisk, ChiSq, CellChi2, Exact, and CMH1. This argument should be of this form:
tables = "var1*var2 / table1 option1 table2 table3"}
}
\description{
(limited) Emulation of SAS proc_freq
}
\examples{
tab1 <-  proc_freq(Rhabdo1, "Freq", "Shakes*Rhabdo / ChiSq CellChi2")
tab2 <- proc_freq(Hawkeye, "Freq", "Year*Win_Loss / Measures")
tab3 <- proc_freq(HA, "Freq", "Trt_Grp*MI / RelRisk")
}
