% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relrisk.R
\name{relrisk}
\alias{relrisk}
\title{Relative Risk}
\usage{
relrisk(x, col)
}
\description{
Computes relative risk with confidence intervals

 @param x A contingency table
 @param col Column number in which to compute relative risk
}
