% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhabdo1_data.R
\docType{data}
\name{Rhabdo1}
\alias{Rhabdo1}
\title{Hawkeye Rhabdo Football Study}
\format{An object of class \code{"data.table"}}
\usage{
data(Rhabdo1)
}
\description{
Rhabdo dataset including None, 1_to_3 and 4_to_5
}
\examples{
data(Rhabdo1)
print(Rhabdo1)
tabs <- proc_freq(Rhabdo1, "Freq", "Shakes*Rhabdo / ChiSq CellChi2 Exact")
}
\references{
BIOS:7410 class notes
}
\keyword{datasets}
