% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhabdo2_data.R
\docType{data}
\name{Rhabdo2}
\alias{Rhabdo2}
\title{Hawkeye Rhabdo Football Study}
\format{An object of class \code{"data.table"}}
\usage{
data(Rhabdo2)
}
\description{
Rhabdo dataset including 1_to_3 and 4_to_5
}
\examples{
data(Rhabdo2)
print(Rhabdo2)
tabs <- proc_freq(Rhabdo2, "Freq", "Shakes*Rhabdo / ChiSq CellChi2")
}
\references{
BIOS:7410 class notes
}
\keyword{datasets}
