% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhabdo3_data.R
\docType{data}
\name{Rhabdo3}
\alias{Rhabdo3}
\title{Hawkeye Rhabdo Football Study}
\format{An object of class \code{"data.table"}}
\usage{
data(Rhabdo3)
}
\description{
Rhabdo dataset including None and Any
}
\examples{
data(Rhabdo3)
print(Rhabdo3)
tabs <- proc_freq(Rhabdo3, "Freq", "Shakes*Rhabdo / ChiSq CellChi2")
}
\references{
BIOS:7410 class notes
}
\keyword{datasets}
