% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{sassyr}
\alias{sassyr}
\alias{sassyr-package}
\title{Simple SAS-inspired tools for exploring data frames}
\description{
Simple SAS-inspired tools for exploring data frames
}
\details{
The main functions in this package are:
\describe{
  \item{\code{\link{content}}}{brief summary of an entire data frame}
  \item{\code{\link{freq}}}{(cross-)tabulations for data frames and vectors}
}
}
\author{
Written and maintained by Daniel Muenz <dmuenz@umich.edu>
}

